Many changes on the bitsets length.
Their seemed to be, imo, some vast confusion concerning the length of the different bitsets.
The password is 12 bits long, the hash 64 and the fingerprint 24.

The DES algo takes the password (12 bits), then expands it to 56 bits by adding 44 0's
(and then 8 parity bits to sum up to 64 bits). It produces an 8 bytes long hash, that
is 64 bits. We only keep the 24 LSB in order to get the so called fingerprint.

The reduction function will then take this fingerprint and, as its name suggests,
reduce it to a 12 bits long word, just like the password (it's his purpose).

I did not take the time to understand what's going on in the reduction function.
I was going for when I began, but rectifying the bitsets took more time than expected.