Implementation of the DES encryption.
It seems to be working. However, I didn't take the time to review the whole code yet.
I may do it later tonight, or tomorrow.

The principle behind this encryption is the following:
we need two things for the encryption: a message and a key, both 64 bits long.
The first is, in the context of this project, contstant and kind of irrelevant.
The later is based on our 12 bits long password, extended to 64 bits.

The extension works this way:
64 bits is 8 bytes, each LSB of each byte being a parity bit. Since we don't care
about them (we don't verify the integrity of the key), their value doesn't matter.
The 6 first bytes will then initialized to 0, and the two last will hold the key.
Let's begin with the last byte will be formed with a prity bit (0, not important)
at the right most side, then the 7 least significant bits of the key.
The penultimate byte, he, will hold the 5 other bits of the key, its partity bit,
and zeros to fill the rest.

Here is what it looks like:
00000000 00000000 00000000 00000000 00000000 00000000 00b12b11b10b9b8p7 b7b6b5b4b3b2b1p8
bi, being the ith bit of the key, pi the ith parity bit ( = 0 ).