Modifications to use the TestReductionFunction.
The TestReductionFunction class is now static, it was useless to create an object for that.
The colorful reduction functions have been implemented and are used in the
RainbowAttack::reductionFunction method.
The main() has been temporarily modified for the test.
The whole RainbowAttack class has to be refactored to respect the encapsulation.
Whatever the note says in the TestReducitonFunctions.cpp, it IS important that the
code is duplicated. This is not normal.

TODO
----

* Refactor RainbowAttack
* Find a way to avoid the code duplication in the TestReductionFunction class
* Improve the search in the table
* Understand why the test on the reduction functions says there is no collision.